IoC now finds dependencies in templates too
the IoC looks for data-dojo-type and requires these as well. That way
we can use digit widgets in our templates without having to require
them first.

Also I now handle when we require objects that are not defined in
modules corresponding to their names but rather with Clazzy directly.

Fixed so that changing config is clean.