Added inherited function
inherited is not as smooth as dojos since I am not using named
functions and hence have to pass the function name I wish to call.
Since the function is also called from the last prototype I don't know
how deep in the inheritance I am and so I must pass the classname, the
args will soon be removed using caller.