Added Data Models in classes, rewritten IoC and added When
Classes created with Clazzy now supports a data model stored in _model.
You can add to the model using addModel(). Setters and Getters will
first search in the model and then in the object. Main reason was
because I want to be able to easily serialize objects for rest services.

Added debug information in the Deferred and DeferredList to that when
something crashes at least you see where the deferred was created and
can start debugging.

IoC got a major rewrite as it was acting a bit dodge in more complex
cases.

Abstraction Layer got reworked to use the Underscore Lib since it is
almost standard. Underscore is now a requirement for Clazzy.

Added then When functionality to the Deferreds so that we can now say:
when something(), () ->
regardless if something returns a deferred or a value